Dodano sporo komentarzy

+++ ApiLibrary/WeatherFeelsLikeModel.cs
@@ -6,11 +6,27 @@ using System.Threading.Tasks;
 
 namespace ApiLibrary
 {
+    // The WeatherTempModel class containes apparent temperature information
+    /// <summary>
+    /// The <c>WeatherTempModel</c> class.
+    /// Containes apparent temperature information.
+    /// </summary>
     public class WeatherFeelsLikeModel
     {
-        public float Day { get; set; }
-        public float Night { get; set; }
-        public float Eve { get; set; }
-        public float Morn { get; set; }
+        // The day apparent temperature
+        /// <value>Gets and sets the day apparent temperature value.</value>
+        public double Day { get; set; }
+
+        // The Night apparent temperature
+        /// <value>Gets and sets the Night apparent temperature value.</value>
+        public double Night { get; set; }
+
+        // The Evening apparent temperature
+        /// <value>Gets and sets the Evening apparent temperature value.</value>
+        public double Eve { get; set; }
+
+        // The Morning apparent temperature
+        /// <value>Gets and sets the Morning apparent temperature value.</value>
+        public double Morn { get; set; }
     }
 }
